chore: Moving IPublicKeySource and related classes to Jwt
