+++ FirebaseAdmin/FirebaseAdmin/Auth/Jwt/IPublicKeySource.cs
@@ -16,7 +16,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace FirebaseAdmin.Auth
+namespace FirebaseAdmin.Auth.Jwt
 {
     /// <summary>
     /// An object that can be used to retrieve a set of RSA public keys for verifying signatures.
